\name{BIC.regimix}
\alias{BIC.regimix}
\title{Information criterion for a regimix model.}
\description{Returns information criterion for regimix models.
}
\section{Method}{ BIC( object, \dots{})
}
\arguments{
\item{ object}{an object obtained from fitting a region of common profile mixture model. Such as that generated from a call to regimix(qv).}
\item{ ...}{ignored}
}
\value{A numeric scalar giving the Bayesian information criterion.
}
\author{Scott D. Foster}
\keyword{misc}
