\name{logLik.regimix}
\alias{logLik.regimix}
\title{The log likelihood for a regimix model.}
\description{Returns the maximised log likelihood for a regimix model.
}
\section{Method}{ logLik( object, \dots{})
}
\arguments{
\item{ object}{an object obtained from fitting a regitat mixture model. Such as that generated from a call to regimix(qv).}
\item{ ...}{ignored}
}
\value{A numeric scalar giving the maximised log likelihood for the regimix model.
}
\author{Scott D. Foster}
\keyword{misc}
