\name{orderFitted}
\alias{orderFitted}
\alias{orderPost}
\title{Assesses classification error and re-orders parameters for simulated data (orderFitted) or for two fits (orderPost).}
\description{Finds the set of (posterior) group labels that most closely matches the simulated data (orderFitted) or another model fit (orderPost). This function is not intended to be used by the general public -- please treat with care. The only help given is the code itself. This is the ultimate R{} experience.
}
\author{Scott D. Foster}
\keyword{misc}
