\name{print.regimix}
\alias{print.regimix}
\title{Prints a regimix object.}
\description{Prints a regimix object. It may be useful for something.
}
\section{Method}{ print( x, \dots{})
}
\arguments{
\item{ x}{an object obtained from fitting a RCP mixture model. Such as that generated from a call to regimix(qv).}
\item{ ...}{ignored}
}
\details{A list is returned that will be printed on exit, if not assigned to anything. It contains the function call and the estimated coefficients.
}
\author{Scott D. Foster}
\keyword{misc}
