\name{summary.regimix}
\alias{summary.regimix}
\title{Summarises a regimix object.}
\description{A summary from a regimix object. It may be useful for something.
}
\section{Method}{ summary( object, \dots{})
}
\arguments{
\item{ object}{a object obtained from fitting a RCP (for region of common profile) mixture model. Such as that generated from a call to regimix(qv).}
\item{ ...}{ignored}
}
\details{ A table is printed that contains the coefficient values, their standard errors, and their z-statistic. The second and thrid columns may be unreliable for some parameters.
}
\author{Scott D. Foster}
\keyword{misc}
